backend/requirements.txt
flask==2.0.1
python-docx==0.8.11
PyMuPDF==1.19.1
spacy==3.2.0
flask-cors==3.0.10
pytesseract==0.3.10
Pillow==9.5.0
